fix: address CR review comments on DCQL and VPFormatSupported
- ClaimValue.IntValue -> LongValue to prevent Long/Number truncation
- ClaimValue.from throws OpenID4VPExceptions instead of IllegalArgumentException
- Rename misleading param isCredentialSetsAvailable -> isClaimSetsAvailable
- CBOR unwrapCborValue uses toLong() instead of toInt() for integer types
- SdJwtVcFormatSupported.toAlgValuesSupported returns kbJwtAlgValues only (Swift parity)
- DcqlEvaluator: stop mislabeling matched claims as CLAIM_UNAVAILABLE

+++ kotlin/openID4VP/src/commonMain/kotlin/io/mosip/openID4VP/evaluator/dcql/DCQLUtils.kt
@@ -343,8 +343,8 @@ private fun extractStringFromCborMap(map: co.nstant.`in`.cbor.model.Map, key: St
 private fun unwrapCborValue(item: DataItem): Any? {
     return when (item) {
         is UnicodeString -> item.string
-        is UnsignedInteger -> item.value.toInt()
-        is NegativeInteger -> item.value.toInt()
+        is UnsignedInteger -> item.value.toLong()
+        is NegativeInteger -> item.value.toLong()
         is DoublePrecisionFloat -> item.value
         is SinglePrecisionFloat -> item.value.toDouble()
         is HalfPrecisionFloat -> item.value.toDouble()
